requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
av==14.3.0
bitsandbytes==0.45.5
black==25.1.0
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/ultralytics/CLIP.git@d19bfc14081d768afe1e4379609444c472f4b38e
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
decord==0.6.0
-e git+ssh://git@github.com/facebookresearch/detectron2.git@9604f5995cc628619f0e4fd913453b4d7d61db3f#egg=detectron2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dotenv==0.9.9
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flash-attn==2.3.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gpustat==1.1.1
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.30.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iopath==0.1.9
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.23
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langgraph==0.3.25
langgraph-checkpoint==2.0.24
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
langsmith==0.3.26
lap==0.5.12
latex2mathml==3.77.0
lazy_loader==0.4
-e git+ssh://git@github.com/fansunqi/LLaVA.git@c9d4e168a7625e742d6880bebb4baadecb0c306d#egg=llava
-e git+ssh://git@github.com/lvis-dataset/lvis-api.git@7d7f07def11da91f8b2710ce352c62a78fd5a7ad#egg=lvis
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mobileclip @ git+https://github.com/apple/ml-mobileclip.git@1140b8d197e4ed7d56b3a92216ded98bb1c2ac87
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.34.1
networkx==3.2.1
ninja==1.11.1.4
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open_clip_torch==2.32.0
openai==1.75.0
opencv-python==4.11.0.86
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
-e git+ssh://git@github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0#egg=panopticapi
pathspec==0.12.1
peft==0.3.0
pillow==10.4.0
platformdirs==4.3.7
portalocker==3.1.1
propcache==0.3.1
protobuf==3.20.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pycocotools==2.0.8
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.5
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.1.99
shapely==2.0.7
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
submitit==1.5.2
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2024.8.30
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.19.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.40.1
triton==2.1.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
ultralytics==8.3.103
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.34.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.19.0
zipp==3.21.0
zstandard==0.23.0
